hx71x: Make query protocol easier to work with
Currently the MSB of the sample is encoded in a way which makes the data
slightly tricky to work with; with the current uses of it there is no
problem, as the two flags in the most significant bits are exclusive of
each other, and require a sample value of 0, so testing for values is easy
enough.

In order to more easily pack channel information into the MSB later on,
this changes the protocol to never send the MSB of the sample.

This also changes the config parameter from "gain_channel" to
"gain_channel0", partly to prepare for multi-channel reporting, but
mostly to ensure we break protocol compatibility in case anyone ever
tries to run this revision.